avs-system-flask-app/requirements.txt
bidict==0.22.0
cffi==1.15.0
click==8.1.0
colorama==0.4.4
cryptography==36.0.2
Flask==2.1.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.3.0
python-engineio==4.3.1
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.1.0
